+++ backend/spring-boot-ecommerce-website/src/main/java/com/ecommerce/ecommercewebsite/entity/OrderedProduct.java
@@ -18,7 +18,7 @@ public class OrderedProduct {
     private Long order_id;
 
     @Column(name = "product_id")
-    private int product_id;
+    private Long product_id;
 
     @Column(name = "quantity")
     private int quantity;
